J_Scrapping/requirements.txt
selenium
tqdm
transformers
nltk
mysql-connector-python